Modify a workflow to see if it appears
This is an experiment to figure out why git-for-windows/git workflows
don't appear in the Actions tab of my fork (and don't run), even
though running GHA workflows is enabled in my fork's settings, and
the button to enable workflows from an upstream repository in a fork
does not appear my fork's Actions tab.

This is not meant for merging upstream.